[Feat 1390] client panel and table done, still have to create the popup and table functionalities like dnd and removal

+++ tm/tm.service/src/main/java/org/squashtest/csp/tm/internal/repository/CustomFieldBindingDao.java
@@ -37,6 +37,6 @@ public interface CustomFieldBindingDao extends CustomCustomFieldBindingDao{
 	
 	void persist(CustomFieldBinding binding);
 	
-	int countAllForProjectAndEntity(long projectId, BindableEntity boundEntity);
+	Long countAllForProjectAndEntity(long projectId, BindableEntity boundEntity);
 	
 }
